added link to all news, created checks for controllers in admin, modi…

+++ Inchoo/NewsWithComments/Controller/Adminhtml/Comments/Delete.php
@@ -21,9 +21,11 @@ class Delete extends Action
 
     public function execute()
     {
-        $ids = $this->_request->getParam('selected');
-        $message = $this->commentsRepository->deleteComments($ids);
-        $this->messageManager->addNoticeMessage($message);
+        if ($ids = $this->_request->getParam('selected')) {
+            $message = $this->commentsRepository->deleteComments($ids);
+            $this->messageManager->addNoticeMessage($message);
+            return $this->_redirect('comments/comments');
+        }
         return $this->_redirect('comments/comments');
     }
 }
